Finished Version 1 of Timerist!

+++ document_editor.py
@@ -15,26 +15,33 @@ class EditWindow(QtWidgets.QMainWindow):
         self.layout = QHBoxLayout()
         self.layout2 = QVBoxLayout()
         self.widget = QtWidgets.QWidget(self)
-        self.editor_settings = load_editor_settings(f"users/{self.user}/editor_settings.json")
+        try:
+            self.editor_settings = load_editor_settings(f"users/{self.user}/editor_settings.json")
+        except:
+            self.editor_settings = 0
         self.textEdit = QtWidgets.QTextEdit(self.widget)
         self.textEdit.setGeometry(QtCore.QRect(20, 20, 400, 350))
         self.textEdit.setObjectName("textEdit")
+        self.font = QFont("Times", 12)
+        self.textEdit.setFont(self.font)
+        self.textEdit.setText(text)
 
-        if "'background-color'" in self.editor_settings:
-            color = self.editor_settings["'background-color'"]
-            ev = eval(color)
-            sheet = f"background-color: rgba{ev};"
-            self.textEdit.setStyleSheet(sheet)
-        if "'background-image'" in self.editor_settings:
-            img = self.editor_settings["'background-image'"]
-            sheet = f"background-image: url({img});"
-            self.textEdit.setStyleSheet(sheet)
-        if "'save-on-close'" in self.editor_settings:
-            ev = eval(self.editor_settings["'save-on-close'"])
-            if ev == 'True':
-                self.toggle_close_saving(True)
-            elif ev == 'False':
-                self.toggle_close_saving(False)
+        if type(self.editor_settings) != int:
+            if "'background-color'" in self.editor_settings:
+                color = self.editor_settings["'background-color'"]
+                ev = eval(color)
+                sheet = f"background-color: rgba{ev};"
+                self.textEdit.setStyleSheet(sheet)
+            if "'background-image'" in self.editor_settings:
+                img = self.editor_settings["'background-image'"]
+                sheet = f"background-image: url({img});background-attachment: fixed;"
+                self.textEdit.setStyleSheet(sheet)
+            if "'save-on-close'" in self.editor_settings:
+                ev = eval(self.editor_settings["'save-on-close'"])
+                if ev == 'True':
+                    self.toggle_close_saving(True)
+                elif ev == 'False':
+                    self.toggle_close_saving(False)
 
         self.editor_bg_color_clicked = False
         self.editor_bg_image_clicked = False
@@ -42,10 +49,6 @@ class EditWindow(QtWidgets.QMainWindow):
         self.editor_save_on_close = True
         self.bg_image_filename = None
         self.bg_color = None  
-        self.font = QFont("Times", 12)
-        self.textEdit.setFont(self.font)
-        self.textEdit.setText(text)
-        self.textEdit.setLineWrapMode(QtWidgets.QTextEdit.NoWrap)
         self.tool_btn_size = QtCore.QSize(35, 35)
         self.pushButtonCreate = QtWidgets.QToolButton(self.widget)
         self.pushButtonCreate.setIcon(QtGui.QIcon("images/add.png"))
@@ -543,12 +546,6 @@ class EditWindow(QtWidgets.QMainWindow):
         self.settings_win.resize(500, 400)
         self.settings_win.setWindowIcon(QtGui.QIcon("images/settings.png"))
         self.settings_win.setWindowTitle("Editor Settings")
-        self.edit_from_json = QAction()
-        self.edit_from_json.setIcon(QIcon("images/gears.png"))
-        self.edit_from_json.setToolTip("Open Settings In JSON")
-        self.edit_from_json.triggered.connect(self.open_json)
-        tlb = QToolBar()
-        tlb.addAction(self.edit_from_json)
 
         label_font = QtGui.QFont()
         label_font.setPointSize(20)
@@ -567,7 +564,6 @@ class EditWindow(QtWidgets.QMainWindow):
         tabs.addTab(apperance_tab, "Apperance")
         tabs.addTab(preferances_tab, "Preferances")
 
-        tab_layout.addWidget(tlb)
         tab_layout.addWidget(tabs)
 
         apperance_tab_layout = QVBoxLayout()
@@ -586,15 +582,16 @@ class EditWindow(QtWidgets.QMainWindow):
 
         
         self.selected_bg_widget = QtWidgets.QLabel()
-        if not "'background-color'" in self.editor_settings:
-            self.selected_bg_widget.setText("(255,255,255)")
-            if not "'background-image'" in self.editor_settings:
+        if type(self.editor_settings) != int:
+            if not "'background-color'" in self.editor_settings:
                 self.selected_bg_widget.setText("(255,255,255)")
+                if not "'background-image'" in self.editor_settings:
+                    self.selected_bg_widget.setText("(255,255,255)")
+                else:
+                    self.selected_bg_widget.setText(eval(self.editor_settings["'background-image'"]))       
             else:
-                self.selected_bg_widget.setText(eval(self.editor_settings["'background-image'"]))       
-        else:
-            self.selected_bg_widget.setText(eval(self.editor_settings["'background-color'"]))
-        #theme_field_options_widget.stateChanged.connect(self.settings_change_theme)
+                self.selected_bg_widget.setText(eval(self.editor_settings["'background-color'"]))
+            #theme_field_options_widget.stateChanged.connect(self.settings_change_theme)
 
         background_field_option1 = QtWidgets.QPushButton("Color")
         background_field_option1.clicked.connect(self.change_editor_bg_color)
@@ -633,13 +630,14 @@ class EditWindow(QtWidgets.QMainWindow):
 
         self.show_save_on_close_toggle = AnimatedToggle(checked_color="#36d1d1")
         self.show_save_on_close_toggle.stateChanged.connect(lambda x: self.toggle_close_saving(True) if x else self.toggle_close_saving(False))
-        if not "'save-on-close'" in self.editor_settings:
-            self.show_save_on_close_toggle.setChecked(True)
-        elif "'save-on-close'" in self.editor_settings:
-            if eval(self.editor_settings["'save-on-close'"]) == 'True':
+        if type(self.editor_settings) != int:
+            if not "'save-on-close'" in self.editor_settings:
                 self.show_save_on_close_toggle.setChecked(True)
-            elif eval(self.editor_settings["'save-on-close'"]) == 'False':
-                self.show_save_on_close_toggle.setChecked(False)
+            elif "'save-on-close'" in self.editor_settings:
+                if eval(self.editor_settings["'save-on-close'"]) == 'True':
+                    self.show_save_on_close_toggle.setChecked(True)
+                elif eval(self.editor_settings["'save-on-close'"]) == 'False':
+                    self.show_save_on_close_toggle.setChecked(False)
 
 
         preferances_form_layout.addRow(show_save_on_close_label, self.show_save_on_close_toggle)
@@ -658,45 +656,6 @@ class EditWindow(QtWidgets.QMainWindow):
         self.settings_win.setLayout(tab_layout)
         self.settings_win.show()
 
-    def jsonOk(self):
-        data = self.JSONtext.text()
-        try:
-            json.loads(data)
-            file=open(f"users/{self.user}/editor_settings.json", "w").close()
-            with open(f"users/{self.user}/editor_settings.json", "w") as f:
-                f.write(data)
-                f.close()
-        except:
-            QtWidgets.QMessageBox.warning(self, "Invalid", "Invalid Settings Script Or Syntax Error!")
-
-    def jsonNo(self):
-        self.win.destroy()
-
-    def open_json(self):
-        self.win = QtWidgets.QMainWindow(self)
-        self.win.resize(500, 350)
-        self.win.setWindowTitle("Editor Settings (JSON)")
-        self.win.setWindowIcon(QIcon("images/settings.png"))
-        layout = QVBoxLayout()
-        widget = QtWidgets.QWidget()
-        data = open(f"users/{self.user}/editor_settings.json", "r").read()
-        self.JSONtext = SimpleJSONEditor(widget)
-        self.JSONtext.setText(data)
-        save = QPushButton("Save")
-        save.setShortcut("Ctrl+S")
-        save.clicked.connect(self.jsonOk)
-        cancel = QPushButton("Cancel")
-        cancel.clicked.connect(self.jsonNo)
-        hl = QHBoxLayout()
-        hl.addWidget(save, 50)
-        hl.addWidget(cancel, 50)
-        wid = QWidget()
-        wid.setLayout(hl)
-        layout.addWidget(self.JSONtext)
-        layout.addWidget(wid)
-        widget.setLayout(layout)
-        self.win.setCentralWidget(widget)
-        self.win.show()
 
     def toggle_close_saving(self, x):
         self.toggle_close_saving_clicked = True
@@ -723,29 +682,49 @@ class EditWindow(QtWidgets.QMainWindow):
         try:
             file = QtWidgets.QFileDialog.getOpenFileName(self, "Insert Image", ".", "PNG (*.png);;JPG (*.jpg);;BMP (*.bmp)")
             self.bg_image_filename = file[0]
-            self.textEdit.setStyleSheet(f"background-image: url({self.bg_image_filename});")
+            self.textEdit.setStyleSheet(f"background-image: url({self.bg_image_filename});background-attachment: fixed;")
             self.selected_bg_widget.setText(str(self.bg_image_filename))
         except:
             QtWidgets.QMessageBox.critical(self, "Fatal!", "Could not set selected image as background.") 
 
     def preferancesOk(self):
         if self.toggle_close_saving_clicked == True:
-            previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
-            previous_data["'save-on-close'"] = f"'{self.editor_save_on_close}'"
-            save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            try:
+                previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
+            except:
+                previous_data = 0
+            if type(previous_data) != int:
+                previous_data["'save-on-close'"] = f"'{self.editor_save_on_close}'"
+                save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            else:
+                save_editor_settings(f"users/{self.user}/editor_settings.json", {"'save-on-close'":f"'{self.editor_save_on_close}'"})
+
 
     def preferancesNo(self):
         self.settings_win.destroy()
 
     def apperanceOk(self):
         if self.editor_bg_color_clicked == True:
-            previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
-            previous_data["'background-color'"] = f"'{self.bg_color}'"
-            save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            try:
+                previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
+            except:
+                previous_data = 0
+            if type(previous_data) != int:
+                previous_data["'background-color'"] = f"'{self.bg_color}'"
+                save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            else:
+                save_editor_settings(f"users/{self.user}/editor_settings.json", {"'background-color'":f"'{self.bg_color}'"})
         elif self.editor_bg_image_clicked == True:
-            previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
-            previous_data["'background-image'"] = f"'{self.bg_image_filename}'"
-            save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            try:
+                previous_data = load_editor_settings(f"users/{self.user}/editor_settings.json")
+            except:
+                previous_data = 0
+            if type(previous_data) != int:
+                previous_data["'background-image'"] = f"'{self.bg_image_filename}'"
+                save_editor_settings(f"users/{self.user}/editor_settings.json", previous_data)
+            else:
+                save_editor_settings(f"users/{self.user}/editor_settings.json", {"'background-image'":f"'{self.bg_image_filename}'"})
+
 
     def apperanceNo(self):
         self.settings_win.destroy()
@@ -768,22 +747,23 @@ class EditWindow(QtWidgets.QMainWindow):
             elif self.editor_save_on_close == False:
                 pass
         else:
-            if "'save-on-close'" in self.editor_settings:
-                if eval(self.editor_settings["'save-on-close'"]) == 'True':
-                    msg_save = QtWidgets.QMessageBox.question(self, "Save Changes", "Would you like to save your changes?", QtWidgets.QMessageBox.Yes|QtWidgets.QMessageBox.No)
-                    if msg_save == QtWidgets.QMessageBox.Yes:
-                        try:
-                            file = open(f"users/{self.user}/database/{self.title}", 'w', encoding='utf-8').close()
-                            with open(f"users/{self.user}/database/{self.title}", "w", encoding='utf-8') as f:
-                                f.write(self.textEdit.toHtml())
-                                f.close()
-                            QtWidgets.QMessageBox.information(self, "Saved!", f"Your changes were saved successfully.")
-                        except:
-                            QtWidgets.QMessageBox.warning(self, "Saving Error", "Please open an existing document to save your changes.")
+            if type(self.editor_settings) != int:
+                if "'save-on-close'" in self.editor_settings:
+                    if eval(self.editor_settings["'save-on-close'"]) == 'True':
+                        msg_save = QtWidgets.QMessageBox.question(self, "Save Changes", "Would you like to save your changes?", QtWidgets.QMessageBox.Yes|QtWidgets.QMessageBox.No)
+                        if msg_save == QtWidgets.QMessageBox.Yes:
+                            try:
+                                file = open(f"users/{self.user}/database/{self.title}", 'w', encoding='utf-8').close()
+                                with open(f"users/{self.user}/database/{self.title}", "w", encoding='utf-8') as f:
+                                    f.write(self.textEdit.toHtml())
+                                    f.close()
+                                QtWidgets.QMessageBox.information(self, "Saved!", f"Your changes were saved successfully.")
+                            except:
+                                QtWidgets.QMessageBox.warning(self, "Saving Error", "Please open an existing document to save your changes.")
+                        else:
+                            pass
                     else:
                         pass
-                else:
-                    pass
 
     def createdocument(self):
         if self.database != None:
